--- norns-job-core/src/main/java/com/gourd/norns/core/MainApp.java
@@ -1,10 +1,0 @@
-package com.gourd.norns.core;
-
-/**
- * @author Li.Wei by 2019/8/30
- */
-public class MainApp {
-    public static void main(String[] args) {
-        System.out.println("norns-job");
-    }
-}
