+++ norns-job-core/src/main/java/io/github/nornslab/norns/core/MainApp.java
@@ -1,0 +1,10 @@
+package io.github.nornslab.norns.core;
+
+/**
+ * @author Li.Wei by 2019/8/30
+ */
+public class MainApp {
+    public static void main(String[] args) {
+        System.out.println("norns-job");
+    }
+}
