--- norns-job-core/src/main/scala/com/gourd/norns/core/Context.scala
@@ -1,13 +1,0 @@
-package com.gourd.norns.core
-
-import com.gourd.norns.core.utils.Logging
-
-/** 上下文内容
-  *
-  * @author Li.Wei by 2019/8/29
-  */
-trait Context extends Logging with AutoCloseable {
-  def name: String = this.getClass.getCanonicalName
-
-  override def close(): Unit = {}
-}
